requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
langchain
deeplake
streamlit
python-dotenv
# openai
sentence_transformers
langchain-community
